+++ app/src/main/kotlin/com/simplemobiletools/smsmessenger/models/ThreadError.kt
@@ -1,3 +1,3 @@
 package com.simplemobiletools.smsmessenger.models
 
-data class ThreadError(val messageID: Int) : ThreadItem()
+data class ThreadError(val messageID: Long) : ThreadItem()
